(1) 회원가입시 널값 오류 해결 => [원인] input 태그에 disabled 속성 때문에 발생 => [해결] disabled 에서 readonly 로 바꿔서 해결함
(2) 최초 프로필변경시 적용안되던 현상 해결
=> [원인] 기존 이미지가 없을경우 경로를 못찾기때문에 발생했었음
=> [해결] if 문으로 유저DTO의 imgPath 가 null 이 아닐때만 실행되도록 수정함

(3) 채팅 작성시 hjm 작성자의  sad 메시지 반복되는 현상 해결
=> [원인] 채팅 매퍼에 채팅방번호를 직접 숫자로 기재했었음 (하드코딩)
=> [해결] #{} 을 이용해서 변수로 바꿈

(4) 채팅시 상대방과 프로필 사진이 번갈아가며 출력되는 현상 해결
=> [원인] 로직 설계에 문제가 있었음
=> [해결] 우측에는 항상 본인이,  좌측에는 내가 아닌 사람의 정보의 사진만 갖고오도록 함

(5) 프로필 사진을 못찾을경우 발생하는 에러에 대한 예외처리 추가  ( InvalidPathException )

(6) resources/js/common/profileImg.js 및 tradeImg.js 파일 추가
 .. 프로필 이미지 or 중고 거래 게시판 이미지 파일을 못찾을때에 대한 예외처리도 추가함
 .. 적용대상
  => 내 대화방 목록 및 대화방
 => 메인 게시판 작성자 프로필 및 댓글 작성자 프로필
 => 공지사항 댓글 작성자 프로필
 => 내 정보 프로필
 => 중고 거래 게시판 미리보기(목록, 수정)

(7) 채팅 일부 기능 및 디자인 변경
 => 채팅 보낸 시간이 말풍선 밑에 나오도록 변경함

+++ src/main/java/com/lgy/smile/controller/UserController.java
@@ -2,6 +2,7 @@ package com.lgy.smile.controller;
 
 import java.io.File;
 import java.nio.file.Files;
+import java.nio.file.InvalidPathException;
 import java.nio.file.NoSuchFileException;
 import java.util.HashMap;
 
@@ -406,8 +407,12 @@ public class UserController {
 		
 		UserDto user = devUtils.getUserInfo(session);
 		
-		File deleteFile = new File(user.getImgPath()); // 기존 프로필이미지 경로+파일명을 기준으로 새로운 파일 생성해서		
-		if ( deleteFile.exists() == true ) { deleteFile.delete();	} // 기존 이미지 파일이 존재하면 삭제
+		// 기존 이미지가 존재한다면, 해당 파일을 찾아서 삭제한다
+		if ( user.getImgPath() != null ) {
+			File deleteFile = new File(user.getImgPath()); // 삭제할 파일의 위치로 이동해서		
+			if ( deleteFile.exists() == true ) { deleteFile.delete();	} // 삭제할 파일이 존재한다면 삭 -제 해버림			
+		}
+		
 		user.setImgPath(newProfilePath);				// 새 프로필이미지 경로+파일명을 setter 로 DTO에 설정하고
 		session.setAttribute("userInfo", user);			// 그걸 다시 session 에 넣어주기
 		
@@ -437,6 +442,8 @@ public class UserController {
 			
 		} catch (NoSuchFileException ne) {
 			log.info("★ 해당 경로에 프로필 사진이 없어서 에러 발생");
+		} catch (InvalidPathException ipe) {
+			log.info("★ 프로필 사진 요청하는 경로 상태가 ..? ");
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
